Bumped version to 3.0.6 because of the following changes -Added a new BlockFile (MultiBlockFile) that can grow dynamically without having to specifiy a maximum number of blocks -Implemented a new MappedBitSet that is backed by a bytebuffer instead of a long array -Replaced java.util.BitSet with MappedBitSet in block files that keeps track of records in a bitset.

+++ src/main/java/org/mellowtech/core/io/impl/BlockMapper.java
@@ -103,7 +103,7 @@ public class BlockMapper {
       removed = true;
     }
     if (removed) {
-      int region = blockSize * blocksToMap;
+      long region =  ((long) blockSize) * blocksToMap;
       fc.truncate(offset + ((blockNo + 1) * region));
     }
   }
